UsersServiceWeb.AuthController (users_service v0.1.0)

This controller manages all of the requests related to authentication

Functions

login(conn, map)

It receives the email and password from the email and returns a new pair of tokens as cookies if the authentication was correct

logout(conn, params)

It returns a pair of empty tokens as cookies

refresh(conn, params)

It generates a new pair of tokens if the user that made the request exists and returns them as cookies